Refactor "compiles single file" tests
