+++ package.json
@@ -59,6 +59,7 @@
   "devDependencies": {
     "gulp": "^3.8.11",
     "gulp-sourcemaps": "^1.2.7",
-    "mocha": "*"
+    "mocha": "*",
+    "vinyl-file": "^1.2.1"
   }
 }
